apps/briefs/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bertopic==0.17.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cssselect==1.3.0
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gliclass==0.1.10
google-auth==2.38.0
google-genai==1.10.0
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hdbscan==0.8.40
hpack==4.1.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
json_repair==0.41.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.2
Mako==1.3.9
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.2
multimethod==1.12
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.0
notebook_shim==0.2.4
numba==0.61.0
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.72.0
openpyxl==3.1.5
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-profiling==3.6.6
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
postgrest==1.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
pyquery==2.0.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
readtime==3.0.0
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.2
triton==3.2.0
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.3.0
visions==0.8.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
wordcloud==1.9.4
yarl==1.19.0
ydata-profiling==4.16.1
